fmcjesdadc1: Update project because of HDL update
Updated the project as a consequence of the HDl project update. The current
modification takes into consideration the fact that the data path for the
ADC data has changed. The modification is that there is now a single
instance of the JESD TPL, the cpack and the ADC data dma.

Signed-off-by: Dan Hotoleanu <dan.hotoleanu@analog.com>